flatpak: Better determine if an app can be copied to USB
This commit modifies the logic that decides whether or not to show a
"Copy to USB" button on an app's details page. Currently it just checks
if the app has a collection ID set. After this commit, it will check if
the app's remote has a collection ID set, if the runtime remote has a
collection ID set, and if the app uses flatpak's "extra data" feature.
This should help prevent people from trying to copy apps to USB drives
that can't be distributed that way. This is especially important since
at the moment the "eos-runtimes" remote doesn't have a collection ID set
by default, and some apps still use the legacy runtimes from there.
Additionally, some popular apps like Spotify and Skype use the "extra
data" feature.

I think it's more robust to check if the remote has a collection ID set
(for both the app and runtime) rather than checking the
"flatpak::CollectionId" property, because in the future it's possible a
change could be introduced to set that property from the app's metadata
rather than the remote's configuration, which would break the logic
because it's the remote configuration that determines if the app can be
copied. Additionally, runtime GsApp objects do not seem to have the
collection ID property set.

https://phabricator.endlessm.com/T23651